Renaming.
datakit-gh => datakit-gh-hooks
datakit-github-bridge => datakit-gh-bridge
